testing/ui/requirements.txt
Adafruit-Blinka==7.2.0
adafruit-circuitpython-busdevice==5.1.8
adafruit-circuitpython-mpu6050==1.1.11
adafruit-circuitpython-register==1.9.8
adafruit-circuitpython-typing==1.4.0
Adafruit-PlatformDetect==3.22.1
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
bidict==0.22.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pathfinding==1.0.1
pyftdi==0.54.0
pyserial==3.5
python-engineio==4.3.1
python-socketio==5.5.2
pyusb==1.2.1
yarl==1.7.2
